amigo-oculto-litestar/src/pyproject.toml
[project]
name = "amigo-oculto-litestar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "litestar[standard]>=2.13.0",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "python-jose>=3.3.0",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = [
    "ruff>=0.8.3",
]
